crdx: move stuff to own files
