+++ packages/crdx/src/graph/index.ts
@@ -1,16 +1,23 @@
 ﻿export * from './append.js'
-export * from './graph.js'
+export * from './children.js'
 export * from './concurrency.js'
 export * from './createGraph.js'
-export * from './children.js'
 export * from './decrypt.js'
-export * from './headsAreEqual.js'
-export * from './getSequence.js'
+export * from './getEncryptedLinks.js'
+export * from './getHashes.js'
+export * from './getHead.js'
+export * from './getLink.js'
 export * from './getParentMap.js'
+export * from './getParents.js'
+export * from './getPredecessors.js'
+export * from './getRoot.js'
+export * from './getSequence.js'
+export * from './getSuccessors.js'
+export * from './headsAreEqual.js'
+export * from './isPredecessor.js'
+export * from './isSuccessor.js'
 export * from './merge.js'
-export * from './predecessors.js'
 export * from './redactGraph.js'
 export * from './serialize.js'
-export * from './successors.js'
 export * from './topoSort.js'
 export * from './types.js'
